+++ src/main/java/appeng/client/gui/implementations/GuiInterfaceTerminal.java
@@ -62,6 +62,7 @@ import java.util.List;
 import java.util.*;
 
 import static appeng.client.render.BlockPosHighlighter.hilightBlock;
+import static appeng.client.render.BlockPosHighlighter.turnPlayerTowards;
 import static appeng.helpers.ItemStackHelper.stackFromNBT;
 
 
@@ -350,6 +351,7 @@ public class GuiInterfaceTerminal extends AEBaseGui {
                 }
             } else {
                 hilightBlock(blockPos, System.currentTimeMillis() + 500 * BlockPosUtils.getDistance(blockPos, blockPos2), playerDim);
+                turnPlayerTowards(blockPos);
                 mc.player.sendStatusMessage(PlayerMessages.InterfaceHighlighted.get(blockPos.getX(), blockPos.getY(), blockPos.getZ()), false);
             }
             mc.player.closeScreen();